+++ Sources/SuperwallKit/Paywall/Presentation/Internal Presentation/Operators/CheckPaywallPresentable.swift
@@ -75,7 +75,7 @@ extension Superwall {
       )
 
       let error = InternalPresentationLogic.presentationError(
-        domain: "SWPresentationError",
+        domain: "SWKPresentationError",
         code: 103,
         title: "No UIViewController to present paywall on",
         value: "This usually happens when you call this method before a window was made key and visible."